getting tests to work partially complete
